+++ src/dumbDB.ts
@@ -2,7 +2,7 @@ export interface User {
     id?: string,
     username: string,
     age: number,
-    hobbies: string[] | string | [],
+    hobbies: string[],
 }
 
 export const users: User[] = [];
