+++ tsconfig.json
@@ -3,7 +3,8 @@
     "target": "es6",
     "outDir": "./dist/",
     "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true
   },
   "exclude": [
     "node_modules",
